Changed serialRefNum's type to UInt16, UInt not being defined in codewarrior.

+++ src/Palm/PalmGlue/palm_lib.c
@@ -1,6 +1,6 @@
 #include "palm_lib.h"
 
-static UInt serialRefNum;
+static UInt16 serialRefNum;
 
 void printSerial(char *mesg) {
 	Err errP, err;
